hiver_validation::custom

Trait FieldProvider

pub trait FieldProvider:
    Send
    + Sync
    + 'static {
    // Required method
    fn get_field_value(&self, field: &str) -> Option<String>;
}

The value type for FieldMatchValidator: a map-like accessor. FieldMatchValidator 的值类型：类似 map 的访问器。

Implement this trait on your struct so that FieldMatchValidator can read the two fields by name. 在你的结构体上实现此 trait，以便 FieldMatchValidator 可以按名称读取两个字段。

Required Methods

fn get_field_value(&self, field: &str) -> Option<String>

Return the string representation of a field by name. 按名称返回字段的字符串表示。